Bugfix: Check for version strings on bower package files
cdnizer now more specifically checks for a version string in `bower` or `.bower` files, for the case where there is a `bower.json` with no version, but a `.bower.json` with a version.

Also, as a side effect, now correctly errors if there is no version in the bower file.